requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.22
sqlalchemy>=1.4
psycopg2-binary>=2.9
python-dateutil>=2.8
python-dotenv>=1.0
httpx[http2]>=0.24
